Habit Tracker Widget - Refactor Dashboard, Journal and WeekScheduleBar, merge duplicated CSS (#221)
* Fix habit views

* Add a unique key for each child in a list and add code comments

* Refactoring code and merge duplicated CSS

* Add missing line at EOF to HabitTracker.module.css
