+++ telos-frontend/src/components/habit-tracker/DashboardHabitTracker.jsx
@@ -14,21 +14,25 @@ import {
   FormControlLabel,
 } from '@material-ui/core';
 import WeekScheduleBar from './WeekScheduleBarHabitTracker';
-import habitStyle from './DashboardHabitTracker.module.css';
+import habitStyle from './HabitTracker.module.css';
 
+// This react component draws a Material-UI Habit Tracker in Journal View for the user to note their habituals
 const DashboardHabitTracker = () => {
+  // The functions that update and save the changes
   const [habitTitle, setTitle] = useState('Habit1');
   const [tempTitle, setTempTitle] = useState(habitTitle);
   const [habitDes, setDes] = useState('Habit description, Do Task X, Y many times per week');
   const [tempDes, setTempDes] = useState(habitDes);
-
   const [weeks, setWeeks] = useState('1');
   const [startDate, setStartDate] = useState('2021-01-01');
   const [endType, setEndType] = useState('Never');
   const [endDate, setEndDate] = useState('');
 
+  // If the add button of the main widget is clicked, isEdit will become true then unhides the pop-out window.
+  // If the save or close buttons of the pop-out window is clicked, isEdit will become false then hides the pop-out window.
   const [isEdit, setEditStatus] = useState(false);
 
+  // The following functions are for saving the drafts
   const handleChangeOnName = (event) => {
     setTempTitle(event.target.value);
   };
@@ -53,14 +57,16 @@ const DashboardHabitTracker = () => {
     setEndType(event.target.value);
   };
 
+  // Save the changes and render
   const saveChange = () => {
     setTitle(tempTitle);
     setDes(tempDes);
   };
 
   return (
+    // Main window of the widgets
     <div>
-      <div className={habitStyle.container}>
+      <div className={habitStyle.cardContainer}>
         <div className={habitStyle.HabitTitle}>
           <h2>{habitTitle}</h2>
           <span className={habitStyle.AddIcon}>
@@ -74,6 +80,7 @@ const DashboardHabitTracker = () => {
         </div>
         <WeekScheduleBar />
       </div>
+      {/* Pop out window of the widget */}
       <Dialog
         open={isEdit}
         onClose={() => {
